Fix types.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@122 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/wordhash.h
@@ -18,7 +18,7 @@ typedef struct wh_alloc_node {
 
 typedef struct wh_alloc_str {
   char * buf;
-  int avail;
+  size_t avail;
   int used;
   struct wh_alloc_str *next;
 } wh_alloc_str;
